+++ src/reporters/lang/en.js
@@ -407,6 +407,21 @@ const messages = {
   verboseUpgradeUnlocking: 'Unlocking $0 in the lockfile.',
   folderMissing: "Directory $0 doesn't exist",
   mutexPortBusy: 'Cannot use the network mutex on port $0. It is probably used by another app.',
+
+  auditRunning: 'Auditing packages',
+  auditSummary: '$0 vulnerabilities found - Packages audited: $1',
+  auditSummarySeverity: 'Severity:',
+  auditCritical: '$0 Critical',
+  auditHigh: '$0 High',
+  auditModerate: '$0 Moderate',
+  auditLow: '$0 Low',
+  auditInfo: '$0 Info',
+  auditResolveCommand: '# Run $0 to resolve $1 $2',
+  auditSemverMajorChange: 'SEMVER WARNING: Recommended action is a potentially breaking change',
+  auditManualReview:
+    'Manual Review\nSome vulnerabilities require your attention to resolve\n\nVisit https://go.npm.me/audit-guide for additional guidance',
+  auditRunAuditForDetails: 'Security audit found potential problems. Run "yarn audit" for additional details.',
+  auditOffline: 'Skipping audit. Security audit cannot be performed in offline mode.',
 };
 
 export type LanguageKeys = $Keys<typeof messages>;